feat: example for implement data table with compound components using tanstack react table
description:
create an example for implementation DataTable component with compound components to provide a flexible and reusable table structure. the TableWithCompound component demonstrates the usage of this new DataTable along with its integrated features:
- compound components: the DataTable is structured using compound components (DataTable.SectionToolBar, DataTable.SectionContent, DataTable.SectionPagination) to allow for greater customization and control over the table layout and functionality.
- global filter: integrated DataTableGlobalFilter for filtering table data.
- view options: included DataTableViewOptions for managing table view settings.
- pagination: implemented DataTablePagination for paginating large datasets.
the TableWithCompound component consumes dummy dependency topic data and utilizes the defined column definitions to render the table. this implementation showcases the practical application of the compound component structure and integrated features.

* dependencies:
- @tanstack/react-table